delete non-active admins

+++ src/main/java/by/matsuk/shop/controller/command/impl/admin/DeleteAdminCommand.java
@@ -2,6 +2,7 @@ package by.matsuk.shop.controller.command.impl.admin;
 
 import by.matsuk.shop.controller.Router;
 import by.matsuk.shop.controller.command.Command;
+import by.matsuk.shop.entity.User;
 import by.matsuk.shop.exception.CommandException;
 import by.matsuk.shop.exception.ServiceException;
 import by.matsuk.shop.model.service.UserService;
@@ -11,6 +12,7 @@ import jakarta.servlet.http.HttpSession;
 
 import static by.matsuk.shop.controller.Parameter.USER_ID;
 import static by.matsuk.shop.controller.SessionAttribute.CURRENT_PAGE;
+import static by.matsuk.shop.controller.SessionAttribute.USER;
 
 /**
  * The type Delete user command.
@@ -21,12 +23,21 @@ public class DeleteAdminCommand implements Command {
     @Override
     public Router execute(HttpServletRequest request) throws CommandException {
         try {
-            long userId = Long.parseLong(request.getParameter(USER_ID));
-            Router router = new Router();
-            service.deleteAdmin(userId);
             HttpSession session = request.getSession();
-            String page = (String) session.getAttribute(CURRENT_PAGE);
-            router.setCurrentPage(page);
+            User user = (User) session.getAttribute(USER);
+            User.UserState userState = user.getState();
+            Router router = new Router();
+
+            if (userState != User.UserState.ACTIVE) {
+                long userId = Long.parseLong(request.getParameter(USER_ID));
+                service.deleteAdmin(userId);
+                //HttpSession session = request.getSession();
+                String page = (String) session.getAttribute(CURRENT_PAGE);
+                router.setCurrentPage(page);
+            } else {
+                String page = (String) session.getAttribute(CURRENT_PAGE);
+                router.setCurrentPage(page);
+            }
             return router;
         } catch (ServiceException | NumberFormatException e) {
             throw new CommandException("Exception in a DeleteUserCommand class ", e);
